resolved circular reference and formatted files
