+++ .gitignore
@@ -12,3 +12,8 @@ python3.11
 pip
 pip3
 pip3.11
+classes.dot
+classes.png
+packages.dot
+packages.png
+
